Separate DataTier and CacheTier into separate modules
At least temporarily, makes their fields public. This is in order
to record the metadata only, so this may change.
