+++ src/engine/strat_engine/backstore/mod.rs
@@ -6,7 +6,9 @@
 mod backstore;
 mod blockdev;
 mod blockdevmgr;
+mod cache_tier;
 mod cleanup;
+mod data_tier;
 pub mod device;
 mod metadata;
 mod range_alloc;
